Add `remove_listeners` method on ulabel
Added the remove_listeners method so that it is possible to remove all of ULabel's event listeners without reloading the page.
Because ULabel instantiates event listeners in multiple ways, this was achieved in multiple ways.
First every jquery event listener was added to a .ulabel namespace so that they can all be removed by one jquery .off method.
Then an AbortController was added to ulabel and passed as an option into every vanilla js event listener, so that they can be aborted with an abort method.
Finally, a reference to every ResizeObserver is stored on ULabel instances so that they can be looped through and terminated.